+++ tests/integration/test_integration_generic.py
@@ -24,6 +24,10 @@ class TestIntegrationGeneric(IntegrationTestCase):
             cls.start_index = os.environ.get("START_INDEX")
         else:
             cls.start_index = None
+        if os.environ.get("START_INDEX_FULL_HISTORY"):
+            cls.start_index_full_history = os.environ.get("START_INDEX_FULL_HISTORY")
+        else:
+            cls.start_index_full_history = None
 
     def test_get_order_history(self):
         # WHEN
@@ -37,7 +41,7 @@ class TestIntegrationGeneric(IntegrationTestCase):
     def test_get_order_history_full_details(self):
         # WHEN
         orders = self.amazon_orders.get_order_history(year=self.year,
-                                                      start_index=self.start_index,
+                                                      start_index=self.start_index_full_history,
                                                       full_details=True)
 
         # THEN